requirements.txt
fastapi[standard]==0.128.0
uvicorn==0.40.0
sqlalchemy==2.0.45
psycopg2-binary==2.9.11
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.2.2
bcrypt==4.0.1
pydantic==2.12.5
python-multipart==0.0.21